+++ modules/debts_api/app/sidekiq/debts_api/v0/digital_dispute_submission_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'debts_api/v0/financial_status_report_service'
+
+module DebtsApi
+  class V0::DigitalDisputeJob
+    def perform(submission_id, user_uuid)
+  end
+end
