+++ modules/debts_api/lib/debts_api/v0/digital_dispute_submission_service.rb
@@ -23,14 +23,13 @@ module DebtsApi
         validate_files_present
         validate_files
 
+        DebtsApi::V0::DigitalDisputeJob.perform(user, base_64_files)
         send_to_dmc
         in_progress_form&.destroy
         {
           success: true,
           message: 'Digital dispute submission received successfully'
         }
-      rescue => e
-        failure_result(e)
       end
 
       private
@@ -59,53 +58,53 @@ module DebtsApi
         InProgressForm.form_for_user('DISPUTE-DEBT', @user)
       end
 
-      def validate_files_present
-        if files.blank? || !files.is_a?(Array) || files.empty?
-          raise NoFilesProvidedError,
-                'at least one file is required'
-        end
-      end
+      # def validate_files_present
+      #   if files.blank? || !files.is_a?(Array) || files.empty?
+      #     raise NoFilesProvidedError,
+      #           'at least one file is required'
+      #   end
+      # end
+      #
+      # def validate_files
+      #   errors = []
+      #
+      #   files.each_with_index do |file, index|
+      #     file_index = index + 1
+      #
+      #     errors << "File #{file_index} must be a PDF" unless file.content_type == ACCEPTED_CONTENT_TYPE
+      #
+      #     errors << "File #{file_index} is too large (maximum is 1MB)" if file.size > MAX_FILE_SIZE
+      #   end
+      #
+      #   raise InvalidFileTypeError, errors.join(', ') if errors.any?
+      # end
+      #
+      # def sanitize_filename(filename)
+      #   name = File.basename(filename)
+      #   name = name.tr(':', '_')
+      #   name.gsub(/[.](?=.*[.])/, '')
+      # end
 
-      def validate_files
-        errors = []
-
-        files.each_with_index do |file, index|
-          file_index = index + 1
-
-          errors << "File #{file_index} must be a PDF" unless file.content_type == ACCEPTED_CONTENT_TYPE
-
-          errors << "File #{file_index} is too large (maximum is 1MB)" if file.size > MAX_FILE_SIZE
-        end
-
-        raise InvalidFileTypeError, errors.join(', ') if errors.any?
-      end
-
-      def sanitize_filename(filename)
-        name = File.basename(filename)
-        name = name.tr(':', '_')
-        name.gsub(/[.](?=.*[.])/, '')
-      end
-
-      def failure_result(error)
-        Rails.logger.error("DigitalDisputeSubmissionService error: #{error.message}\n#{error.backtrace&.join("\n")}")
-        case error
-        when NoFilesProvidedError
-          {
-            success: false,
-            errors: { files: [error.message] }
-          }
-        when InvalidFileTypeError
-          {
-            success: false,
-            errors: { files: error.message.split(', ') }
-          }
-        else
-          {
-            success: false,
-            errors: { base: ['An error occurred processing your submission'] }
-          }
-        end
-      end
+      # def failure_result(error)
+      #   Rails.logger.error("DigitalDisputeSubmissionService error: #{error.message}\n#{error.backtrace&.join("\n")}")
+      #   case error
+      #   when NoFilesProvidedError
+      #     {
+      #       success: false,
+      #       errors: { files: [error.message] }
+      #     }
+      #   when InvalidFileTypeError
+      #     {
+      #       success: false,
+      #       errors: { files: error.message.split(', ') }
+      #     }
+      #   else
+      #     {
+      #       success: false,
+      #       errors: { base: ['An error occurred processing your submission'] }
+      #     }
+      #   end
+      # end
     end
   end
 end
